tts_service/pyproject.toml
[tool.setuptools.packages.find]
where = ["."]

[project]
name = "sandy-tts-service"
version = "0.1.0"
description = "Local Qwen TTS service for the Sandy voice spike"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.2",
    "pydantic>=2.12.5",
    "python-dotenv>=1.0.0",
    "qwen-tts>=0.1.1",
    "uvicorn>=0.40.0",
]

[project.scripts]
sandy-tts-service = "tts_service.__main__:main"
